fusion_emotion_model_marketplace_package/requirements.txt
torch==2.1.0
torchvision
torchaudio
transformers==4.40.1
opencv-python-headless==4.10.0.82
numpy
pandas==2.2.2
scikit-learn==1.5.0
PyYAML
Pillow
pytorch-lightning
torchmetrics
fastapi==0.110.2
uvicorn[standard]==0.29.0
boto3==1.34.102
matplotlib==3.8.4
facenet-pytorch==2.5.2
sounddevice==0.4.6
